fix(libcgroup): fix disable_oom_killer in cgroup v1 (#3090)
* fix(libcgroup): fix disable_oom_killer in cgroup v1

The intended behavior is to enable the OOM killer when
disable_oom_killer is not set, which is the common case. However, since
0 means enable and 1 means disable, the if condition is currently
reversed.

This is a really critical issue, as it will cause processes in the
container to hang.

Signed-off-by: xujihui1985 <xujihui1985@gmail.com>

* refactor: rename function disable_oom_killer

Rename set_disable_oom_kill to set_oom_control to make the function name
more meaningful.

Signed-off-by: xujihui1985 <xujihui1985@gmail.com>

* fix: cargo fmt

Signed-off-by: xujihui1985 <xujihui1985@gmail.com>

---------

Signed-off-by: xujihui1985 <xujihui1985@gmail.com>